Add files via upload
All the inputs are numbers and get added to a list and when a 0 is added it ends the program. I have to print the numbers that are in an even position and the list in the reverse order. Then split the main list in two other lists. From the first list print  the numbers except the first and the last. And from the second list print the max and the min.